llm_experiment/requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
alpaca-eval==0.3.1
-e git+https://github.com/tatsu-lab/alpaca_farm.git@b17d605a360fc2936cbe6ee89e81d7270b440903#egg=alpaca_farm
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli
cattrs==23.1.2
certifi
cffi
charset-normalizer 
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.0
crcmod==1.7
cryptography
cuda-python==12.3.0
cycler==0.12.1
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.11.1
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
docstring-parser==0.15
einops==0.7.0
evaluate==0.4.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock
fire==0.5.0
flash-attn==2.4.2
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gmpy2
hjson==3.1.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna
ipdb==0.13.13
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft
mkl-random
mkl-service==2.4.0
mpmath
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx
ninja==1.11.1.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy
openai==0.28.1
optimum==1.13.3
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
peft==0.6.2
pexpect==4.8.0
Pillow
pipreqs==0.4.13
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.0
pycparser
pydantic==1.10.13
Pygments==2.16.1
pyOpenSSL
pyparsing==3.1.1
PySocks
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-rapidjson==1.13
pytz==2023.3.post1
PyYAML
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
ray==2.7.1
referencing==0.30.2
regex==2023.10.3
requests
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
shtab==1.6.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
sympy
tabulate==0.9.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchtyping==0.1.4
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.36.2
triton==2.1.0
tritonclient==2.39.0
-e git+https://github.com/Asap7772/trl.git@1e4933cd6005f3e3107862e417132682894e7673#egg=trl
-e git+https://github.com/CarperAI/trlx.git@78c7faa600105a8f90a318a24bc65f8af9a9053a#egg=trlx
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions
tyro==0.5.15
tzdata==2023.3
uri-template==1.3.0
urllib3 
wandb==0.16.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.2
